+++ templates/element/funding/workshopWithFundingAllowed.php
@@ -11,14 +11,22 @@ echo '<div class="workshop-wrapper">';
     echo '<div class="table">';
 
         $classes = ['button'];
+        $buttonHref = $this->Html->urlFundingsEdit($workshop->uid);
         if ($workshop->funding_created_by_different_owner) {
             $classes[] = 'disabled';
+            $buttonHref = 'javascript:void(0);';
+        }
+        $isSubmitted = $workshop->funding_exists && $workshop->workshop_funding->is_submitted;
+        if ($isSubmitted) {
+            $classes[] = 'disabled';
+            $buttonHref = 'javascript:void(0);';
         }
         echo $this->Html->link(
             $workshop->funding_exists ? 'Förderantrag bearbeiten' : 'Förderantrag erstellen',
-            $workshop->funding_created_by_different_owner ? 'javascript:void(0);' : $this->Html->urlFundingsEdit($workshop->uid),
+            $buttonHref,
             [
                 'class' => implode(' ', $classes),
+                'disabled' => $isSubmitted,
             ],
         );
 
@@ -28,9 +36,14 @@ echo '<div class="workshop-wrapper">';
             }
             echo $this->element('funding/owner', ['funding' => $workshop->workshop_funding]);
             echo $this->element('funding/orgaTeam', ['orgaTeam' => $workshop->orga_team]);
-            echo $this->element('funding/activityProof', ['workshop' => $workshop]);
-            echo $this->element('funding/freistellungsbescheid', ['workshop' => $workshop]);
-            if ($workshop->funding_exists) {
+            if (!$isSubmitted) {
+                echo $this->element('funding/activityProof', ['workshop' => $workshop]);
+                echo $this->element('funding/freistellungsbescheid', ['workshop' => $workshop]);
+            } else {
+                echo $this->element('funding/submitInfo', ['funding' => $workshop->workshop_funding]);
+                echo $this->element('funding/zuwendungsbestaetigungInfo', ['funding' => $workshop->workshop_funding]);
+           }
+            if ($workshop->funding_exists && !$workshop->workshop_funding->is_submitted) {
                 echo $this->element('funding/delete', ['funding' => $workshop->workshop_funding]);
             }
         echo '</div>';
